requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudscraper==1.2.71
exceptiongroup==1.1.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
numpy==1.21.6
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pluggy==1.0.0
pyparsing==3.0.9
PySimpleGUI==4.60.5
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.2
zipp==3.15.0
